+++ cuda_implementation/sigmoid.h
@@ -5,6 +5,6 @@
 #define EULER_NUMBER_F 2.71828182846
 #define EULER_NUMBER_L 2.71828182845904523536
 
-__global__  double sigmoidl( double n);
+__global__  float sigmoidl( float n);
 
 #endif
